feature(frontend): Ability to disable registration
