+++ client/src/App.jsx
@@ -1,64 +1,11 @@
-import { createBrowserRouter, RouterProvider, Navigate, Outlet } from 'react-router-dom';
-import Root from './routes/Root';
-import Chat from './routes/Chat';
-import Search from './routes/Search';
+import { RouterProvider } from 'react-router-dom';
 import { ScreenshotProvider } from './utils/screenshotContext.jsx';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { Login, Registration, RequestPasswordReset, ResetPassword } from './components/Auth';
-import { AuthContextProvider } from './hooks/AuthContext';
 import { RecoilRoot } from 'recoil';
 import { QueryClient, QueryClientProvider, QueryCache } from '@tanstack/react-query';
 import { ThemeProvider } from './hooks/ThemeContext';
 import { useApiErrorBoundary } from './hooks/ApiErrorBoundaryContext';
-import ApiErrorWatcher from './components/Auth/ApiErrorWatcher';
-
-const AuthLayout = () => (
-  <AuthContextProvider>
-    <Outlet />
-    <ApiErrorWatcher />
-  </AuthContextProvider>
-);
-const router = createBrowserRouter([
-  {
-    path: 'register',
-    element: <Registration />
-  },
-  {
-    path: 'forgot-password',
-    element: <RequestPasswordReset />
-  },
-  {
-    path: 'reset-password',
-    element: <ResetPassword />
-  },
-  {
-    element: <AuthLayout />,
-    children: [
-      {
-        path: 'login',
-        element: <Login />
-      },
-      {
-        path: '/',
-        element: <Root />,
-        children: [
-          {
-            index: true,
-            element: <Navigate to="/chat/new" replace={true} />
-          },
-          {
-            path: 'chat/:conversationId?',
-            element: <Chat />
-          },
-          {
-            path: 'search/:query?',
-            element: <Search />
-          }
-        ]
-      }
-    ]
-  }
-]);
+import router from './routes';
 
 const App = () => {
   const { setError } = useApiErrorBoundary();
